requirements.txt
Django==2.2.2
django-cors-headers==3.0.2
django-froala-editor==2.9.5
djangorestframework==3.9.4
